requirements.txt
asyncpg==0.30.0
python-dotenv==1.1.0
fastapi==0.115.13
httpx==0.28.1
numpy==2.1.3
pandas==2.2.3
pydantic==2.10.3
sqlalchemy==2.0.39
uvicorn==0.32.1
